requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
slowapi==0.1.8
email-validator==2.1.0.post1
pymysql==1.1.0
cryptography==42.0.2
alembic==1.13.1
itsdangerous==2.1.2
starlette==0.36.3
bcrypt==4.1.2
typing-extensions>=4.8.0 